resolve conversation
