+++ adminapi/src/main/java/io/minio/admin/messages/info/ErasureSetInfo.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package io.minio.admin.messages;
+package io.minio.admin.messages.info;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
