+++ adminapi/src/main/java/io/minio/admin/messages/info/TimedAction.java
@@ -15,51 +15,37 @@
  * limitations under the License.
  */
 
-package io.minio.admin.messages;
+package io.minio.admin.messages.info;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.math.BigDecimal;
 
 /**
- * MemStats is strip down version of runtime.MemStats containing memory stats of MinIO server.
+ * TimedAction contains a number of actions and their accumulated duration in nanoseconds.
  *
- * @see <a href= "https://github.com/minio/madmin-go/blob/main/health.go#L856">health.go</a>
+ * @see <a href= "https://github.com/minio/madmin-go/blob/main/metrics.go#L244">metrics.go</a>
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class MemStats {
-  @JsonProperty("Alloc")
-  private BigDecimal alloc;
+public class TimedAction {
+  @JsonProperty("count")
+  private BigDecimal count;
 
-  @JsonProperty("TotalAlloc")
-  private BigDecimal totalAlloc;
+  @JsonProperty("acc_time_ns")
+  private BigDecimal accTime;
 
-  @JsonProperty("Mallocs")
-  private BigDecimal mallocs;
+  @JsonProperty("bytes")
+  private BigDecimal bytes;
 
-  @JsonProperty("Frees")
-  private BigDecimal frees;
-
-  @JsonProperty("HeapAlloc")
-  private BigDecimal heapAlloc;
-
-  public BigDecimal alloc() {
-    return alloc;
+  public BigDecimal count() {
+    return count;
   }
 
-  public BigDecimal totalAlloc() {
-    return totalAlloc;
+  public BigDecimal accTime() {
+    return accTime;
   }
 
-  public BigDecimal mallocs() {
-    return mallocs;
-  }
-
-  public BigDecimal frees() {
-    return frees;
-  }
-
-  public BigDecimal heapAlloc() {
-    return heapAlloc;
+  public BigDecimal bytes() {
+    return bytes;
   }
 }
